// src/types/settings.ts
export interface PersistedSettings {
  songBackground: string
  songBodyBackground: string
  bibleBackground: string
  songTitleColor: string
  songLyricsColor: string
  bibleTitleColor: string
  bibleLyricsColor: string
  songTitleFont: string
  songTitleSize: number
  songTitleWeight: string
  songLyricsFont: string
  songLyricsSize: number
  songLyricsWeight: string
  bibleTitleFont: string
  bibleTitleSize: number
  bibleTitleWeight: string
  bibleLyricsFont: string
  bibleLyricsSize: number
  bibleLyricsWeight: string
  projectionMode: "default" | "subtitle"
}